[WIP][flink] Decouple tag Auto Creation logic
