+++ paimon-core/src/main/java/org/apache/paimon/tag/TagAutoCreation.java
@@ -20,57 +20,34 @@ package org.apache.paimon.tag;
 
 import org.apache.paimon.CoreOptions;
 import org.apache.paimon.Snapshot;
-import org.apache.paimon.data.Timestamp;
 import org.apache.paimon.operation.TagDeletion;
+import org.apache.paimon.tag.extractor.ProcessTimeExtractor;
+import org.apache.paimon.tag.extractor.TimeExtractor;
+import org.apache.paimon.tag.extractor.WatermarkExtractor;
+import org.apache.paimon.tag.period.DailyTagPeriodHandler;
+import org.apache.paimon.tag.period.HourlyTagPeriodHandler;
+import org.apache.paimon.tag.period.TagPeriodHandler;
+import org.apache.paimon.tag.period.TwoHoursTagPeriodHandler;
 import org.apache.paimon.utils.SnapshotManager;
 import org.apache.paimon.utils.TagManager;
 
-import javax.annotation.Nullable;
-
 import java.time.Duration;
-import java.time.Instant;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
-import java.time.format.ResolverStyle;
-import java.time.format.SignStyle;
 import java.util.Optional;
 import java.util.SortedMap;
 
-import static java.time.temporal.ChronoField.DAY_OF_MONTH;
-import static java.time.temporal.ChronoField.HOUR_OF_DAY;
-import static java.time.temporal.ChronoField.MONTH_OF_YEAR;
-import static java.time.temporal.ChronoField.YEAR;
 import static org.apache.paimon.Snapshot.FIRST_SNAPSHOT_ID;
 import static org.apache.paimon.shade.guava30.com.google.common.base.MoreObjects.firstNonNull;
-import static org.apache.paimon.utils.Preconditions.checkArgument;
 
-/** A manager to create tags automatically. */
+/**
+ * The TagAutoCreation class automates the process of tagging snapshots according to various rules
+ * and configurations, such as time extraction modes, tagging periods, and retention policies.
+ *
+ * <p>This class manages the lifecycle of tags, creates new tags based on time criteria, and
+ * optionally deletes old tags based on retention policies.
+ */
 public class TagAutoCreation {
-
-    private static final DateTimeFormatter HOUR_FORMATTER =
-            new DateTimeFormatterBuilder()
-                    .appendValue(YEAR, 1, 10, SignStyle.NORMAL)
-                    .appendLiteral('-')
-                    .appendValue(MONTH_OF_YEAR, 2, 2, SignStyle.NORMAL)
-                    .appendLiteral('-')
-                    .appendValue(DAY_OF_MONTH, 2, 2, SignStyle.NORMAL)
-                    .appendLiteral(" ")
-                    .appendValue(HOUR_OF_DAY, 2, 2, SignStyle.NORMAL)
-                    .toFormatter()
-                    .withResolverStyle(ResolverStyle.LENIENT);
-
-    private static final DateTimeFormatter DAY_FORMATTER =
-            new DateTimeFormatterBuilder()
-                    .appendValue(YEAR, 1, 10, SignStyle.NORMAL)
-                    .appendLiteral('-')
-                    .appendValue(MONTH_OF_YEAR, 2, 2, SignStyle.NORMAL)
-                    .appendLiteral('-')
-                    .appendValue(DAY_OF_MONTH, 2, 2, SignStyle.NORMAL)
-                    .toFormatter()
-                    .withResolverStyle(ResolverStyle.LENIENT);
 
     private final SnapshotManager snapshotManager;
     private final TagManager tagManager;
@@ -79,7 +56,6 @@ public class TagAutoCreation {
     private final TagPeriodHandler periodHandler;
     private final Duration delay;
     private final Integer numRetainedMax;
-
     private LocalDateTime nextTag;
     private long nextSnapshot;
 
@@ -98,18 +74,15 @@ public class TagAutoCreation {
         this.periodHandler = periodHandler;
         this.delay = delay;
         this.numRetainedMax = numRetainedMax;
-
         this.periodHandler.validateDelay(delay);
 
         SortedMap<Snapshot, String> tags = tagManager.tags();
-
         if (tags.isEmpty()) {
             this.nextSnapshot =
                     firstNonNull(snapshotManager.earliestSnapshotId(), FIRST_SNAPSHOT_ID);
         } else {
             Snapshot lastTag = tags.lastKey();
             this.nextSnapshot = lastTag.id() + 1;
-
             LocalDateTime time = periodHandler.tagToTime(tags.get(lastTag));
             this.nextTag = periodHandler.nextTagTime(time);
         }
@@ -118,7 +91,7 @@ public class TagAutoCreation {
     public void run() {
         while (true) {
             if (snapshotManager.snapshotExists(nextSnapshot)) {
-                tryToTag(snapshotManager.snapshot(nextSnapshot));
+                attemptTagCreation(snapshotManager.snapshot(nextSnapshot));
                 nextSnapshot++;
             } else {
                 // avoid snapshot has been expired
@@ -132,7 +105,7 @@ public class TagAutoCreation {
         }
     }
 
-    private void tryToTag(Snapshot snapshot) {
+    private void attemptTagCreation(Snapshot snapshot) {
         Optional<LocalDateTime> timeOptional = timeExtractor.extract(snapshot);
         if (!timeOptional.isPresent()) {
             return;
@@ -167,143 +140,6 @@ public class TagAutoCreation {
         return t1.isAfter(t2) || t1.isEqual(t2);
     }
 
-    private interface TimeExtractor {
-
-        Optional<LocalDateTime> extract(Snapshot snapshot);
-    }
-
-    private static class ProcessTimeExtractor implements TimeExtractor {
-
-        @Override
-        public Optional<LocalDateTime> extract(Snapshot snapshot) {
-            return Optional.of(
-                    Instant.ofEpochMilli(snapshot.timeMillis())
-                            .atZone(ZoneId.systemDefault())
-                            .toLocalDateTime());
-        }
-    }
-
-    private static class WatermarkExtractor implements TimeExtractor {
-
-        private final ZoneId watermarkZoneId;
-
-        private WatermarkExtractor(ZoneId watermarkZoneId) {
-            this.watermarkZoneId = watermarkZoneId;
-        }
-
-        @Override
-        public Optional<LocalDateTime> extract(Snapshot snapshot) {
-            Long watermark = snapshot.watermark();
-            if (watermark == null) {
-                return Optional.empty();
-            }
-
-            return Optional.of(
-                    Instant.ofEpochMilli(watermark).atZone(watermarkZoneId).toLocalDateTime());
-        }
-    }
-
-    private interface TagPeriodHandler {
-
-        void validateDelay(Duration delay);
-
-        LocalDateTime tagToTime(String tag);
-
-        LocalDateTime normalizeToTagTime(LocalDateTime time);
-
-        String timeToTag(LocalDateTime time);
-
-        LocalDateTime nextTagTime(LocalDateTime time);
-    }
-
-    private abstract static class BaseTagPeriodHandler implements TagPeriodHandler {
-
-        protected abstract Duration onePeriod();
-
-        protected abstract DateTimeFormatter formatter();
-
-        @Override
-        public void validateDelay(Duration delay) {
-            checkArgument(onePeriod().compareTo(delay) > 0);
-        }
-
-        @Override
-        public LocalDateTime tagToTime(String tag) {
-            return LocalDateTime.parse(tag, formatter());
-        }
-
-        @Override
-        public LocalDateTime normalizeToTagTime(LocalDateTime time) {
-            long mills = Timestamp.fromLocalDateTime(time).getMillisecond();
-            long periodMills = onePeriod().toMillis();
-            LocalDateTime normalized =
-                    Timestamp.fromEpochMillis((mills / periodMills) * periodMills)
-                            .toLocalDateTime();
-            return normalized.minus(onePeriod());
-        }
-
-        @Override
-        public String timeToTag(LocalDateTime time) {
-            return time.format(formatter());
-        }
-
-        @Override
-        public LocalDateTime nextTagTime(LocalDateTime time) {
-            return time.plus(onePeriod());
-        }
-    }
-
-    private static class HourlyTagPeriodHandler extends BaseTagPeriodHandler {
-
-        private static final Duration ONE_PERIOD = Duration.ofHours(1);
-
-        @Override
-        protected Duration onePeriod() {
-            return ONE_PERIOD;
-        }
-
-        @Override
-        protected DateTimeFormatter formatter() {
-            return HOUR_FORMATTER;
-        }
-    }
-
-    private static class DailyTagPeriodHandler extends BaseTagPeriodHandler {
-
-        private static final Duration ONE_PERIOD = Duration.ofDays(1);
-
-        @Override
-        protected Duration onePeriod() {
-            return ONE_PERIOD;
-        }
-
-        @Override
-        protected DateTimeFormatter formatter() {
-            return DAY_FORMATTER;
-        }
-
-        @Override
-        public LocalDateTime tagToTime(String tag) {
-            return LocalDate.parse(tag, formatter()).atStartOfDay();
-        }
-    }
-
-    private static class TwoHoursTagPeriodHandler extends BaseTagPeriodHandler {
-
-        private static final Duration ONE_PERIOD = Duration.ofHours(2);
-
-        @Override
-        protected Duration onePeriod() {
-            return ONE_PERIOD;
-        }
-
-        @Override
-        protected DateTimeFormatter formatter() {
-            return HOUR_FORMATTER;
-        }
-    }
-
-    @Nullable
     public static TagAutoCreation create(
             CoreOptions options,
             SnapshotManager snapshotManager,
