+++ paimon-core/src/main/java/org/apache/paimon/tag/extractor/TimeExtractor.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.paimon.tag.extractor;
+
+import org.apache.paimon.Snapshot;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+
+/**
+ * The TimeExtractor interface defines the contract for extracting a LocalDateTime from a given
+ * Snapshot object.
+ *
+ * <p>Implementing classes should provide their own mechanism to derive the LocalDateTime from the
+ * Snapshot, and return it wrapped in an Optional. If the extraction is unsuccessful or not
+ * applicable, an empty Optional should be returned.
+ */
+public interface TimeExtractor {
+
+    Optional<LocalDateTime> extract(Snapshot snapshot);
+}
