+++ paimon-core/src/main/java/org/apache/paimon/tag/period/TagPeriodHandler.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.paimon.tag.period;
+
+import java.time.Duration;
+import java.time.LocalDateTime;
+
+/**
+ * The TagPeriodHandler interface defines a set of operations for handling time tags and delays in a
+ * standardized manner. These operations include the validation of delays, conversion between time
+ * tags and LocalDateTime instances, normalization of time to a specific tag time, and finding the
+ * next applicable tag time.
+ *
+ * <p>Implementing classes should define their own logic for these operations.
+ */
+public interface TagPeriodHandler {
+
+    void validateDelay(Duration delay);
+
+    LocalDateTime tagToTime(String tag);
+
+    LocalDateTime normalizeToTagTime(LocalDateTime time);
+
+    String timeToTag(LocalDateTime time);
+
+    LocalDateTime nextTagTime(LocalDateTime time);
+}
